Reconfigured video code, added basic question asking and removed favicon error
