+++ teach_app/teach_app/urls.py
@@ -1,23 +1,9 @@
-"""teach_app URL Configuration
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/3.0/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
 from django.contrib import admin
 from django.urls import path, include
 from django.conf import settings
 from django.conf.urls.static import static
 from teach_app_backend import views
+from django.views.generic import RedirectView
 
 urlpatterns = [
     path('', include('teach_app_backend.urls')),
@@ -38,4 +24,5 @@ urlpatterns = [
     path('student-grade/', views.edit_student_grade, name='student-grade'),
     path('student-feedback/', views.edit_student_feedback, name='student-feedback'),
     path('admin/', admin.site.urls),
+    path('favicon.ico/', RedirectView.as_view(url='/teach_app_frontend/static/frontend/favicon.ico/'))
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
