+++ teach_app/teach_app_frontend/static/frontend/main.js
@@ -3353,7 +3353,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var reac
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);\nfunction _typeof(obj) { \"@babel/helpers - typeof\"; if (typeof Symbol === \"function\" && typeof Symbol.iterator === \"symbol\") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === \"function\" && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }; } return _typeof(obj); }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } return _assertThisInitialized(self); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Date.prototype.toString.call(Reflect.construct(Date, [], function () {})); return true; } catch (e) { return false; } }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\nvar LectureVideo = /*#__PURE__*/function (_Component) {\n  _inherits(LectureVideo, _Component);\n\n  var _super = _createSuper(LectureVideo);\n\n  function LectureVideo(props) {\n    var _this;\n\n    _classCallCheck(this, LectureVideo);\n\n    _this = _super.call(this, props);\n    _this.state = {\n      configuration: {\n        'iceServers': [{\n          'urls': 'stun:eu-turn3.xirsys.com'\n        }]\n      },\n      ident: 'teach',\n      secret: '2b07e9c8-da3b-11ea-ab22-0242ac150002',\n      channel: 'Teach',\n      username: _this.props.userEmail,\n      sendDisabled: true,\n      peerConnections: {}\n    };\n    _this.getToken = _this.getToken.bind(_assertThisInitialized(_this));\n    _this.getSocketServer = _this.getSocketServer.bind(_assertThisInitialized(_this));\n    _this.connectToSocket = _this.connectToSocket.bind(_assertThisInitialized(_this));\n    _this.onSocketMessage = _this.onSocketMessage.bind(_assertThisInitialized(_this));\n    _this.getIceCandidates = _this.getIceCandidates.bind(_assertThisInitialized(_this));\n    _this.onCreateOffer = _this.onCreateOffer.bind(_assertThisInitialized(_this));\n    _this.onCreateAnswer = _this.onCreateAnswer.bind(_assertThisInitialized(_this));\n    _this.onIceCandidate = _this.onIceCandidate.bind(_assertThisInitialized(_this));\n    _this.onDataChannel = _this.onDataChannel.bind(_assertThisInitialized(_this));\n    _this.onMessageChange = _this.onMessageChange.bind(_assertThisInitialized(_this));\n    _this.sendMessage = _this.sendMessage.bind(_assertThisInitialized(_this));\n    _this.onDataMessage = _this.onDataMessage.bind(_assertThisInitialized(_this));\n    _this.onDataChannelOpen = _this.onDataChannelOpen.bind(_assertThisInitialized(_this));\n    _this.callPeer = _this.callPeer.bind(_assertThisInitialized(_this));\n    _this.createNewPeerConnection = _this.createNewPeerConnection.bind(_assertThisInitialized(_this));\n    _this.setDataChannelHandlers = _this.setDataChannelHandlers.bind(_assertThisInitialized(_this));\n    _this.getUsernameByRemoteDescription = _this.getUsernameByRemoteDescription.bind(_assertThisInitialized(_this));\n    _this.getLocalVideoStream = _this.getLocalVideoStream.bind(_assertThisInitialized(_this));\n    return _this;\n  }\n\n  _createClass(LectureVideo, [{\n    key: \"getToken\",\n    value: function getToken() {\n      var _this2 = this;\n\n      var xhr = new XMLHttpRequest();\n\n      xhr.onreadystatechange = function ($evt) {\n        if (xhr.readyState == 4 && xhr.status == 200) {\n          var res = JSON.parse(xhr.responseText);\n          console.log(\"token response: \", res);\n\n          _this2.setState({\n            token: res['v']\n          });\n\n          _this2.getSocketServer();\n        }\n      };\n\n      xhr.open(\"PUT\", \"https://global.xirsys.net/_token/Teach?k=\" + this.state.username, true);\n      xhr.setRequestHeader(\"Authorization\", \"Basic \" + btoa(\"teach:2b07e9c8-da3b-11ea-ab22-0242ac150002\"));\n      xhr.setRequestHeader(\"Content-Type\", \"application/json\");\n      xhr.send();\n    }\n  }, {\n    key: \"getSocketServer\",\n    value: function getSocketServer() {\n      var _this3 = this;\n\n      var xhr = new XMLHttpRequest();\n\n      xhr.onreadystatechange = function ($evt) {\n        if (xhr.readyState == 4 && xhr.status == 200) {\n          var res = JSON.parse(xhr.responseText);\n          console.log(\"server response: \", res);\n\n          _this3.setState({\n            server: res['v']\n          });\n\n          _this3.connectToSocket();\n        }\n      };\n\n      xhr.open(\"GET\", \"https://global.xirsys.net/_host/Teach?type=signal&k=\" + this.state.username, true);\n      xhr.setRequestHeader(\"Authorization\", \"Basic \" + btoa(\"teach:2b07e9c8-da3b-11ea-ab22-0242ac150002\"));\n      xhr.setRequestHeader(\"Content-Type\", \"application/json\");\n      xhr.send();\n    }\n  }, {\n    key: \"connectToSocket\",\n    value: function connectToSocket() {\n      var _this4 = this;\n\n      var serverUrl = this.state.server + \"/v2/\" + this.state.token;\n      var ws = new WebSocket(serverUrl);\n      console.log(ws);\n\n      ws.onopen = function () {\n        ws.addEventListener('message', _this4.onSocketMessage);\n\n        _this4.setState({\n          webSocket: ws\n        });\n\n        _this4.getIceCandidates();\n      };\n    }\n  }, {\n    key: \"onSocketMessage\",\n    value: function onSocketMessage(event) {\n      var _this5 = this;\n\n      var data = JSON.parse(event.data);\n      var f;\n      var desc;\n      var pc;\n      var sender;\n\n      switch (data.m.o) {\n        case \"peers\":\n          var users = data.p.users;\n\n          for (var i = 0; i < users.length; i++) {\n            console.log(users[i] + \" is in the chat\");\n          }\n\n          break;\n\n        case \"peer_connected\":\n          f = data.m.f.split(\"/\");\n          var joining = f[f.length - 1];\n          console.log(joining + \" has joined the chat\");\n          this.callPeer(joining);\n          break;\n\n        case \"message\":\n          switch (data.p.msg.type) {\n            case \"offer\":\n              desc = new RTCSessionDescription(data.p.msg);\n              f = data.m.f.split(\"/\");\n              sender = f[f.length - 1];\n              pc = this.createNewPeerConnection(sender);\n              pc.setRemoteDescription(desc);\n              pc.createAnswer().then(function (d) {\n                return _this5.onCreateAnswer(d, sender);\n              });\n              break;\n\n            case \"answer\":\n              desc = new RTCSessionDescription(data.p.msg);\n              f = data.m.f.split(\"/\");\n              sender = f[f.length - 1];\n              this.state.peerConnections[sender].pc.setRemoteDescription(desc);\n              break;\n\n            case \"candidate\":\n              f = data.m.f.split(\"/\");\n              sender = f[f.length - 1];\n              console.log(\"candidate\", sender, data, this.state.peerConnections);\n              var candidate = new RTCIceCandidate(data.p.msg);\n              this.state.peerConnections[sender].pc.addIceCandidate(candidate);\n              break;\n          }\n\n      }\n    }\n  }, {\n    key: \"getIceCandidates\",\n    value: function getIceCandidates() {\n      var _this6 = this;\n\n      var xhr = new XMLHttpRequest();\n\n      xhr.onreadystatechange = function ($evt) {\n        if (xhr.readyState == 4 && xhr.status == 200) {\n          var res = JSON.parse(xhr.responseText);\n          console.log(\"ice candidates: \", res);\n\n          _this6.setState({\n            iceCandidates: res['v']\n          });\n        }\n      };\n\n      xhr.open(\"PUT\", \"https://global.xirsys.net/_turn/Teach\", true);\n      xhr.setRequestHeader(\"Authorization\", \"Basic \" + btoa(\"teach:2b07e9c8-da3b-11ea-ab22-0242ac150002\"));\n      xhr.setRequestHeader(\"Content-Type\", \"application/json\");\n      xhr.send();\n    }\n  }, {\n    key: \"onCreateOffer\",\n    value: function onCreateOffer(description, peer) {\n      this.state.peerConnections[peer].pc.setLocalDescription(description);\n      var pkt = {\n        t: \"u\",\n        m: {\n          f: \"Teach/\" + this.state.username,\n          o: \"message\",\n          t: peer\n        },\n        p: {\n          msg: description\n        }\n      };\n      this.state.webSocket.send(JSON.stringify(pkt));\n    }\n  }, {\n    key: \"onCreateAnswer\",\n    value: function onCreateAnswer(description, peer) {\n      this.state.peerConnections[peer].pc.setLocalDescription(description);\n      var pkt = {\n        t: \"u\",\n        m: {\n          f: \"Teach/\" + this.state.username,\n          o: \"message\",\n          t: peer\n        },\n        p: {\n          msg: description\n        }\n      };\n      this.state.webSocket.send(JSON.stringify(pkt));\n    }\n  }, {\n    key: \"onIceCandidate\",\n    value: function onIceCandidate(event) {\n      var remoteUsername;\n\n      if (event.target.remoteDescription) {\n        remoteUsername = this.getUsernameByRemoteDescription(event.target.remoteDescription.sdp);\n      }\n\n      ;\n      var candidate = event.candidate;\n\n      if (candidate != null && remoteUsername != null) {\n        var cPkt = {\n          type: \"candidate\",\n          sdpMLineIndex: candidate.sdpMLineIndex,\n          sdpMid: candidate.sdpMid,\n          candidate: candidate.candidate\n        };\n        var pkt = {\n          t: \"u\",\n          m: {\n            f: \"Teach/\" + this.state.username,\n            o: \"message\",\n            t: remoteUsername\n          },\n          p: {\n            msg: cPkt\n          }\n        };\n        this.state.webSocket.send(JSON.stringify(pkt));\n      }\n    }\n  }, {\n    key: \"onDataChannel\",\n    value: function onDataChannel(event) {\n      var dataChannel = event.channel;\n      var keys = Object.keys(this.state.peerConnections);\n      var comp;\n\n      for (var i = 0; i < keys.length; i++) {\n        comp = this.state.peerConnections[keys[i]];\n\n        if (event.currentTarget.localDescription.sdp == comp.pc.localDescription.sdp) {\n          comp.dc = dataChannel;\n        }\n      }\n\n      this.setDataChannelHandlers(dataChannel);\n    }\n  }, {\n    key: \"setDataChannelHandlers\",\n    value: function setDataChannelHandlers(dc) {\n      var _this7 = this;\n\n      dc.onmessage = function (evt) {\n        return _this7.onDataMessage(evt);\n      };\n\n      dc.onopen = function (evt) {\n        return _this7.onDataChannelOpen(evt);\n      };\n    }\n  }, {\n    key: \"onMessageChange\",\n    value: function onMessageChange(event) {\n      this.setState({\n        message: event.target.value\n      });\n    }\n  }, {\n    key: \"onDataChannelOpen\",\n    value: function onDataChannelOpen(event) {\n      this.setState({\n        sendDisabled: false\n      });\n    }\n  }, {\n    key: \"callPeer\",\n    value: function callPeer(peer) {\n      var _this8 = this;\n\n      if (peer !== this.state.username) {\n        var pc = this.createNewPeerConnection(peer);\n        var dataChannel = pc.createDataChannel(\"data\");\n        this.state.peerConnections[peer].dc = dataChannel;\n        this.setDataChannelHandlers(dataChannel);\n        pc.createOffer().then(function (d) {\n          return _this8.onCreateOffer(d, peer);\n        });\n      }\n\n      ;\n    }\n  }, {\n    key: \"sendMessage\",\n    value: function sendMessage() {\n      var message = this.state.message;\n      console.log(\"You said \" + message);\n      var messagePacket = {\n        f: this.state.username,\n        msg: message\n      };\n      var dataChannel;\n      var keys = Object.keys(this.state.peerConnections);\n      var comp;\n\n      for (var i = 0; i < keys.length; i++) {\n        comp = this.state.peerConnections[keys[i]];\n        dataChannel = comp.dc;\n        dataChannel.send(JSON.stringify(messagePacket));\n      }\n    }\n  }, {\n    key: \"onDataMessage\",\n    value: function onDataMessage(event) {\n      var messagePacket = JSON.parse(event.data);\n      console.log(messagePacket.f + \" said: \" + messagePacket.msg);\n    }\n  }, {\n    key: \"createNewPeerConnection\",\n    value: function createNewPeerConnection(username) {\n      var _this9 = this;\n\n      var pc = new RTCPeerConnection(this.state.iceCandidates);\n\n      if (this.props.userType === \"teacher\") {\n        pc.addStream(this.state.localStream);\n      } else {\n        pc.onaddstream = function (event) {\n          return _this9.onAddStream(event);\n        };\n      }\n\n      ;\n\n      pc.ondatachannel = function (event) {\n        return _this9.onDataChannel(event);\n      };\n\n      pc.onicecandidate = function (candidate) {\n        return _this9.onIceCandidate(candidate);\n      };\n\n      pc.oniceconnectionstatechange = function (event) {\n        return console.log(\"iceconnection state \", event);\n      };\n\n      this.state.peerConnections[username] = {\n        pc: pc,\n        dc: null\n      };\n      console.log(\"setting new peer connection: \", this.state.peerConnections);\n      return pc;\n    }\n  }, {\n    key: \"getUsernameByRemoteDescription\",\n    value: function getUsernameByRemoteDescription(sdp) {\n      var keys = Object.keys(this.state.peerConnections);\n      var pc;\n\n      for (var i = 0; i < keys.length; i++) {\n        pc = this.state.peerConnections[keys[i]].pc;\n\n        if (pc.remoteDescription.sdp == sdp) {\n          return keys[i];\n        }\n      }\n\n      return null;\n    }\n  }, {\n    key: \"getLocalVideoStream\",\n    value: function getLocalVideoStream() {\n      var _this10 = this;\n\n      var constraints = {\n        audio: true,\n        video: true\n      };\n      navigator.mediaDevices.getUserMedia(constraints).then(function (stream) {\n        _this10.setState({\n          localStream: stream\n        });\n\n        console.log('Got MediaStream:', stream);\n        var video = document.querySelector('video');\n        video.srcObject = stream;\n\n        video.onloadedmetadata = function () {\n          video.play();\n        };\n      })[\"catch\"](function (error) {\n        console.error('Error accessing media devices.', error);\n      });\n    }\n  }, {\n    key: \"onAddStream\",\n    value: function onAddStream(event) {\n      var video = document.querySelector('video');\n      video.srcObject = event.stream;\n      console.log(\"Received media stream: \" + event.stream);\n\n      video.onloadedmetadata = function () {\n        video.play();\n      };\n    }\n  }, {\n    key: \"componentDidMount\",\n    value: function componentDidMount() {\n      this.getToken();\n\n      if (this.props.userType === \"teacher\") {\n        this.getLocalVideoStream();\n      }\n\n      ;\n    }\n  }, {\n    key: \"render\",\n    value: function render() {\n      return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"input\", {\n        onChange: this.onMessageChange,\n        type: \"text\",\n        name: \"message\",\n        placeholder: \"Message\"\n      }), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"button\", {\n        onClick: this.sendMessage,\n        disabled: this.state.sendDisabled\n      }, \"Send\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"video\", {\n        autoPlay: true,\n        controls: true\n      }));\n    }\n  }]);\n\n  return LectureVideo;\n}(react__WEBPACK_IMPORTED_MODULE_0__[\"Component\"]);\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (LectureVideo);\n\n//# sourceURL=webpack:///./src/components/LectureVideo/LectureVideo.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);\nfunction _typeof(obj) { \"@babel/helpers - typeof\"; if (typeof Symbol === \"function\" && typeof Symbol.iterator === \"symbol\") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === \"function\" && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }; } return _typeof(obj); }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } return _assertThisInitialized(self); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Date.prototype.toString.call(Reflect.construct(Date, [], function () {})); return true; } catch (e) { return false; } }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\nvar LectureVideo = /*#__PURE__*/function (_Component) {\n  _inherits(LectureVideo, _Component);\n\n  var _super = _createSuper(LectureVideo);\n\n  function LectureVideo() {\n    _classCallCheck(this, LectureVideo);\n\n    return _super.apply(this, arguments);\n  }\n\n  _createClass(LectureVideo, [{\n    key: \"componentDidMount\",\n    value: function componentDidMount() {\n      var video = document.querySelector('video');\n      video.srcObject = this.props.lectureStream;\n\n      video.onloadedmetadata = function () {\n        video.play();\n      };\n    }\n  }, {\n    key: \"render\",\n    value: function render() {\n      return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"video\", {\n        autoPlay: true,\n        controls: true\n      }));\n    }\n  }]);\n\n  return LectureVideo;\n}(react__WEBPACK_IMPORTED_MODULE_0__[\"Component\"]);\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (LectureVideo);\n\n//# sourceURL=webpack:///./src/components/LectureVideo/LectureVideo.js?");
 
 /***/ }),
 
@@ -3515,6 +3515,18 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var reac
 /***/ (function(module, exports, __webpack_require__) {
 
 eval("var api = __webpack_require__(/*! ../../../node_modules/style-loader/dist/runtime/injectStylesIntoStyleTag.js */ \"./node_modules/style-loader/dist/runtime/injectStylesIntoStyleTag.js\");\n            var content = __webpack_require__(/*! !../../../node_modules/css-loader/dist/cjs.js??ref--5-1!./StudentSubmission.module.css */ \"./node_modules/css-loader/dist/cjs.js?!./src/components/StudentSubmission/StudentSubmission.module.css\");\n\n            content = content.__esModule ? content.default : content;\n\n            if (typeof content === 'string') {\n              content = [[module.i, content, '']];\n            }\n\nvar options = {};\n\noptions.insert = \"head\";\noptions.singleton = false;\n\nvar update = api(content, options);\n\n\n\nmodule.exports = content.locals || {};\n\n//# sourceURL=webpack:///./src/components/StudentSubmission/StudentSubmission.module.css?");
+
+/***/ }),
+
+/***/ "./src/components/SubmitQuestion/SubmitQuestion.js":
+/*!*********************************************************!*\
+  !*** ./src/components/SubmitQuestion/SubmitQuestion.js ***!
+  \*********************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _UI_Button_Button__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../UI/Button/Button */ \"./src/components/UI/Button/Button.js\");\nfunction _typeof(obj) { \"@babel/helpers - typeof\"; if (typeof Symbol === \"function\" && typeof Symbol.iterator === \"symbol\") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === \"function\" && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }; } return _typeof(obj); }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } return _assertThisInitialized(self); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Date.prototype.toString.call(Reflect.construct(Date, [], function () {})); return true; } catch (e) { return false; } }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\n\nvar SubmitQuestion = /*#__PURE__*/function (_Component) {\n  _inherits(SubmitQuestion, _Component);\n\n  var _super = _createSuper(SubmitQuestion);\n\n  function SubmitQuestion(props) {\n    var _this;\n\n    _classCallCheck(this, SubmitQuestion);\n\n    _this = _super.call(this, props);\n    _this.onQuestionChange = _this.onQuestionChange.bind(_assertThisInitialized(_this));\n    _this.submitQuestion = _this.submitQuestion.bind(_assertThisInitialized(_this));\n    return _this;\n  }\n\n  _createClass(SubmitQuestion, [{\n    key: \"onQuestionChange\",\n    value: function onQuestionChange(event) {\n      this.setState({\n        question: event.target.value\n      });\n    }\n  }, {\n    key: \"submitQuestion\",\n    value: function submitQuestion() {\n      var question = this.state.question;\n      var questionPacket = {\n        f: this.props.username,\n        question: question\n      };\n      var dataChannel;\n      var keys = Object.keys(this.props.peerConnections);\n      var peer;\n\n      for (var i = 0; i < keys.length; i++) {\n        peer = this.props.peerConnections[keys[i]];\n        dataChannel = peer.dc;\n        dataChannel.send(JSON.stringify(questionPacket));\n      }\n    }\n  }, {\n    key: \"render\",\n    value: function render() {\n      return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"input\", {\n        onChange: this.onQuestionChange,\n        type: \"text\",\n        name: \"question\",\n        placeholder: \"Ask a question\"\n      }), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_UI_Button_Button__WEBPACK_IMPORTED_MODULE_1__[\"default\"], {\n        clicked: this.submitQuestion\n      }, \"Submit Question\"));\n    }\n  }]);\n\n  return SubmitQuestion;\n}(react__WEBPACK_IMPORTED_MODULE_0__[\"Component\"]);\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (SubmitQuestion);\n\n//# sourceURL=webpack:///./src/components/SubmitQuestion/SubmitQuestion.js?");
 
 /***/ }),
 
@@ -3735,7 +3747,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var reac
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _components_LectureVideo_LectureVideo__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../../../components/LectureVideo/LectureVideo */ \"./src/components/LectureVideo/LectureVideo.js\");\n\n\n\nvar lecture = function lecture(props) {\n  return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", null, props.lecture, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_components_LectureVideo_LectureVideo__WEBPACK_IMPORTED_MODULE_1__[\"default\"], {\n    userEmail: props.userEmail,\n    userType: props.userType\n  }));\n};\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (lecture);\n\n//# sourceURL=webpack:///./src/containers/Main/Lectures/Lecture/Lecture.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _components_LectureVideo_LectureVideo__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../../../components/LectureVideo/LectureVideo */ \"./src/components/LectureVideo/LectureVideo.js\");\n/* harmony import */ var _components_SubmitQuestion_SubmitQuestion__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../../../components/SubmitQuestion/SubmitQuestion */ \"./src/components/SubmitQuestion/SubmitQuestion.js\");\nfunction _typeof(obj) { \"@babel/helpers - typeof\"; if (typeof Symbol === \"function\" && typeof Symbol.iterator === \"symbol\") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === \"function\" && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }; } return _typeof(obj); }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } return _assertThisInitialized(self); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Date.prototype.toString.call(Reflect.construct(Date, [], function () {})); return true; } catch (e) { return false; } }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\n\n\nvar Lecture = /*#__PURE__*/function (_Component) {\n  _inherits(Lecture, _Component);\n\n  var _super = _createSuper(Lecture);\n\n  function Lecture(props) {\n    var _this;\n\n    _classCallCheck(this, Lecture);\n\n    _this = _super.call(this, props);\n    _this.state = {\n      username: _this.props.userEmail,\n      sendDisabled: true,\n      peerConnections: {}\n    };\n    _this.getToken = _this.getToken.bind(_assertThisInitialized(_this));\n    _this.getSocketServer = _this.getSocketServer.bind(_assertThisInitialized(_this));\n    _this.connectToSocket = _this.connectToSocket.bind(_assertThisInitialized(_this));\n    _this.onSocketMessage = _this.onSocketMessage.bind(_assertThisInitialized(_this));\n    _this.getIceCandidates = _this.getIceCandidates.bind(_assertThisInitialized(_this));\n    _this.onCreateOffer = _this.onCreateOffer.bind(_assertThisInitialized(_this));\n    _this.onCreateAnswer = _this.onCreateAnswer.bind(_assertThisInitialized(_this));\n    _this.onIceCandidate = _this.onIceCandidate.bind(_assertThisInitialized(_this));\n    _this.onDataChannel = _this.onDataChannel.bind(_assertThisInitialized(_this));\n    _this.onMessageChange = _this.onMessageChange.bind(_assertThisInitialized(_this));\n    _this.onDataMessage = _this.onDataMessage.bind(_assertThisInitialized(_this));\n    _this.onDataChannelOpen = _this.onDataChannelOpen.bind(_assertThisInitialized(_this));\n    _this.callPeer = _this.callPeer.bind(_assertThisInitialized(_this));\n    _this.createNewPeerConnection = _this.createNewPeerConnection.bind(_assertThisInitialized(_this));\n    _this.setDataChannelHandlers = _this.setDataChannelHandlers.bind(_assertThisInitialized(_this));\n    _this.getUsernameByRemoteDescription = _this.getUsernameByRemoteDescription.bind(_assertThisInitialized(_this));\n    _this.getLocalVideoStream = _this.getLocalVideoStream.bind(_assertThisInitialized(_this));\n    return _this;\n  }\n\n  _createClass(Lecture, [{\n    key: \"getToken\",\n    value: function getToken() {\n      var _this2 = this;\n\n      var xhr = new XMLHttpRequest();\n\n      xhr.onreadystatechange = function ($evt) {\n        if (xhr.readyState == 4 && xhr.status == 200) {\n          var res = JSON.parse(xhr.responseText);\n          console.log(\"token response: \", res);\n\n          _this2.setState({\n            token: res['v']\n          });\n\n          _this2.getSocketServer();\n        }\n      };\n\n      xhr.open(\"PUT\", \"https://global.xirsys.net/_token/Teach?k=\" + this.state.username, true);\n      xhr.setRequestHeader(\"Authorization\", \"Basic \" + btoa(\"teach:2b07e9c8-da3b-11ea-ab22-0242ac150002\"));\n      xhr.setRequestHeader(\"Content-Type\", \"application/json\");\n      xhr.send();\n    }\n  }, {\n    key: \"getSocketServer\",\n    value: function getSocketServer() {\n      var _this3 = this;\n\n      var xhr = new XMLHttpRequest();\n\n      xhr.onreadystatechange = function ($evt) {\n        if (xhr.readyState == 4 && xhr.status == 200) {\n          var res = JSON.parse(xhr.responseText);\n          console.log(\"server response: \", res);\n\n          _this3.setState({\n            server: res['v']\n          });\n\n          _this3.connectToSocket();\n        }\n      };\n\n      xhr.open(\"GET\", \"https://global.xirsys.net/_host/Teach?type=signal&k=\" + this.state.username, true);\n      xhr.setRequestHeader(\"Authorization\", \"Basic \" + btoa(\"teach:2b07e9c8-da3b-11ea-ab22-0242ac150002\"));\n      xhr.setRequestHeader(\"Content-Type\", \"application/json\");\n      xhr.send();\n    }\n  }, {\n    key: \"connectToSocket\",\n    value: function connectToSocket() {\n      var _this4 = this;\n\n      var serverUrl = this.state.server + \"/v2/\" + this.state.token;\n      var ws = new WebSocket(serverUrl);\n      console.log(ws);\n\n      ws.onopen = function () {\n        ws.addEventListener('message', _this4.onSocketMessage);\n\n        _this4.setState({\n          webSocket: ws\n        });\n\n        _this4.getIceCandidates();\n      };\n    }\n  }, {\n    key: \"onSocketMessage\",\n    value: function onSocketMessage(event) {\n      var _this5 = this;\n\n      var data = JSON.parse(event.data);\n      var f;\n      var desc;\n      var pc;\n      var sender;\n\n      switch (data.m.o) {\n        case \"peers\":\n          var users = data.p.users;\n\n          for (var i = 0; i < users.length; i++) {\n            console.log(users[i] + \" is in the chat\");\n          }\n\n          break;\n\n        case \"peer_connected\":\n          f = data.m.f.split(\"/\");\n          var joining = f[f.length - 1];\n          console.log(joining + \" has joined the chat\");\n          this.callPeer(joining);\n          break;\n\n        case \"message\":\n          switch (data.p.msg.type) {\n            case \"offer\":\n              desc = new RTCSessionDescription(data.p.msg);\n              f = data.m.f.split(\"/\");\n              sender = f[f.length - 1];\n              pc = this.createNewPeerConnection(sender);\n              pc.setRemoteDescription(desc);\n              pc.createAnswer().then(function (d) {\n                return _this5.onCreateAnswer(d, sender);\n              });\n              break;\n\n            case \"answer\":\n              desc = new RTCSessionDescription(data.p.msg);\n              f = data.m.f.split(\"/\");\n              sender = f[f.length - 1];\n              this.state.peerConnections[sender].pc.setRemoteDescription(desc);\n              break;\n\n            case \"candidate\":\n              f = data.m.f.split(\"/\");\n              sender = f[f.length - 1];\n              console.log(\"candidate\", sender, data, this.state.peerConnections);\n              var candidate = new RTCIceCandidate(data.p.msg);\n              this.state.peerConnections[sender].pc.addIceCandidate(candidate);\n              break;\n          }\n\n      }\n    }\n  }, {\n    key: \"getIceCandidates\",\n    value: function getIceCandidates() {\n      var _this6 = this;\n\n      var xhr = new XMLHttpRequest();\n\n      xhr.onreadystatechange = function ($evt) {\n        if (xhr.readyState == 4 && xhr.status == 200) {\n          var res = JSON.parse(xhr.responseText);\n          console.log(\"ice candidates: \", res);\n\n          _this6.setState({\n            iceCandidates: res['v']\n          });\n        }\n      };\n\n      xhr.open(\"PUT\", \"https://global.xirsys.net/_turn/Teach\", true);\n      xhr.setRequestHeader(\"Authorization\", \"Basic \" + btoa(\"teach:2b07e9c8-da3b-11ea-ab22-0242ac150002\"));\n      xhr.setRequestHeader(\"Content-Type\", \"application/json\");\n      xhr.send();\n    }\n  }, {\n    key: \"onCreateOffer\",\n    value: function onCreateOffer(description, peer) {\n      this.state.peerConnections[peer].pc.setLocalDescription(description);\n      var pkt = {\n        t: \"u\",\n        m: {\n          f: \"Teach/\" + this.state.username,\n          o: \"message\",\n          t: peer\n        },\n        p: {\n          msg: description\n        }\n      };\n      this.state.webSocket.send(JSON.stringify(pkt));\n    }\n  }, {\n    key: \"onCreateAnswer\",\n    value: function onCreateAnswer(description, peer) {\n      this.state.peerConnections[peer].pc.setLocalDescription(description);\n      var pkt = {\n        t: \"u\",\n        m: {\n          f: \"Teach/\" + this.state.username,\n          o: \"message\",\n          t: peer\n        },\n        p: {\n          msg: description\n        }\n      };\n      this.state.webSocket.send(JSON.stringify(pkt));\n    }\n  }, {\n    key: \"onIceCandidate\",\n    value: function onIceCandidate(event) {\n      var remoteUsername;\n\n      if (event.target.remoteDescription) {\n        remoteUsername = this.getUsernameByRemoteDescription(event.target.remoteDescription.sdp);\n      }\n\n      ;\n      var candidate = event.candidate;\n\n      if (candidate != null && remoteUsername != null) {\n        var cPkt = {\n          type: \"candidate\",\n          sdpMLineIndex: candidate.sdpMLineIndex,\n          sdpMid: candidate.sdpMid,\n          candidate: candidate.candidate\n        };\n        var pkt = {\n          t: \"u\",\n          m: {\n            f: \"Teach/\" + this.state.username,\n            o: \"message\",\n            t: remoteUsername\n          },\n          p: {\n            msg: cPkt\n          }\n        };\n        this.state.webSocket.send(JSON.stringify(pkt));\n      }\n    }\n  }, {\n    key: \"onDataChannel\",\n    value: function onDataChannel(event) {\n      var dataChannel = event.channel;\n      var keys = Object.keys(this.state.peerConnections);\n      var comp;\n\n      for (var i = 0; i < keys.length; i++) {\n        comp = this.state.peerConnections[keys[i]];\n\n        if (event.currentTarget.localDescription.sdp == comp.pc.localDescription.sdp) {\n          comp.dc = dataChannel;\n        }\n      }\n\n      this.setDataChannelHandlers(dataChannel);\n    }\n  }, {\n    key: \"setDataChannelHandlers\",\n    value: function setDataChannelHandlers(dc) {\n      var _this7 = this;\n\n      dc.onmessage = function (evt) {\n        return _this7.onDataMessage(evt);\n      };\n\n      dc.onopen = function (evt) {\n        return _this7.onDataChannelOpen(evt);\n      };\n    }\n  }, {\n    key: \"onMessageChange\",\n    value: function onMessageChange(event) {\n      this.setState({\n        message: event.target.value\n      });\n    }\n  }, {\n    key: \"onDataChannelOpen\",\n    value: function onDataChannelOpen(event) {\n      this.setState({\n        sendDisabled: false\n      });\n    }\n  }, {\n    key: \"callPeer\",\n    value: function callPeer(peer) {\n      var _this8 = this;\n\n      if (peer !== this.state.username) {\n        var pc = this.createNewPeerConnection(peer);\n        var dataChannel = pc.createDataChannel(\"data\");\n        this.state.peerConnections[peer].dc = dataChannel;\n        this.setDataChannelHandlers(dataChannel);\n        pc.createOffer().then(function (d) {\n          return _this8.onCreateOffer(d, peer);\n        });\n      }\n\n      ;\n    } // sendMessage() {\n    //     const message = this.state.message;\n    //     console.log(\"You said \" + message);\n    //     const messagePacket = {\n    //         f: this.state.username,\n    //         msg: message\n    //     };\n    //     var dataChannel;\n    //     const keys = Object.keys(this.state.peerConnections);\n    //     var comp;\n    //     for(var i = 0; i < keys.length; i++) {\n    //         comp = this.state.peerConnections[keys[i]];\n    //         dataChannel = comp.dc;\n    //         dataChannel.send(JSON.stringify(messagePacket));\n    //     }\n    // }\n\n  }, {\n    key: \"onDataMessage\",\n    value: function onDataMessage(event) {\n      var questionPacket = JSON.parse(event.data);\n      console.log(questionPacket.f + \" asked: \" + questionPacket.question);\n    }\n  }, {\n    key: \"createNewPeerConnection\",\n    value: function createNewPeerConnection(username) {\n      var _this9 = this;\n\n      var pc = new RTCPeerConnection(this.state.iceCandidates);\n\n      if (this.props.userType === \"teacher\") {\n        pc.addStream(this.state.lectureStream);\n      } else {\n        pc.onaddstream = function (event) {\n          return _this9.onAddStream(event);\n        };\n      }\n\n      ;\n\n      pc.ondatachannel = function (event) {\n        return _this9.onDataChannel(event);\n      };\n\n      pc.onicecandidate = function (candidate) {\n        return _this9.onIceCandidate(candidate);\n      };\n\n      pc.oniceconnectionstatechange = function (event) {\n        return console.log(\"iceconnection state \", event);\n      };\n\n      this.state.peerConnections[username] = {\n        pc: pc,\n        dc: null\n      };\n      console.log(\"setting new peer connection: \", this.state.peerConnections);\n      return pc;\n    }\n  }, {\n    key: \"getUsernameByRemoteDescription\",\n    value: function getUsernameByRemoteDescription(sdp) {\n      var keys = Object.keys(this.state.peerConnections);\n      var pc;\n\n      for (var i = 0; i < keys.length; i++) {\n        pc = this.state.peerConnections[keys[i]].pc;\n\n        if (pc.remoteDescription.sdp == sdp) {\n          return keys[i];\n        }\n      }\n\n      return null;\n    }\n  }, {\n    key: \"getLocalVideoStream\",\n    value: function getLocalVideoStream() {\n      var _this10 = this;\n\n      var constraints = {\n        audio: true,\n        video: true\n      };\n      navigator.mediaDevices.getUserMedia(constraints).then(function (stream) {\n        _this10.setState({\n          lectureStream: stream\n        });\n\n        console.log('Got MediaStream:', stream);\n      })[\"catch\"](function (error) {\n        console.error('Error accessing media devices.', error);\n      });\n    }\n  }, {\n    key: \"onAddStream\",\n    value: function onAddStream(event) {\n      this.setState({\n        lectureStream: event.stream\n      });\n      console.log(\"Received media stream: \" + event.stream);\n    }\n  }, {\n    key: \"componentDidMount\",\n    value: function componentDidMount() {\n      this.getToken();\n\n      if (this.props.userType === \"teacher\") {\n        this.getLocalVideoStream();\n      }\n\n      ;\n    }\n  }, {\n    key: \"render\",\n    value: function render() {\n      var lectureVideo;\n\n      if (this.state.lectureStream) {\n        lectureVideo = /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_components_LectureVideo_LectureVideo__WEBPACK_IMPORTED_MODULE_1__[\"default\"], {\n          lectureStream: this.state.lectureStream\n        });\n      }\n\n      var submitQuestion;\n\n      if (this.props.userType === \"student\") {\n        submitQuestion = /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_components_SubmitQuestion_SubmitQuestion__WEBPACK_IMPORTED_MODULE_2__[\"default\"], {\n          peerConnections: this.state.peerConnections,\n          username: this.state.username\n        });\n      }\n\n      return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", null, lectureVideo, submitQuestion);\n    }\n  }]);\n\n  return Lecture;\n}(react__WEBPACK_IMPORTED_MODULE_0__[\"Component\"]);\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Lecture);\n\n//# sourceURL=webpack:///./src/containers/Main/Lectures/Lecture/Lecture.js?");
 
 /***/ }),
